# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for a table top game.",
        "title": "Table Top API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/damage_combo_cards": {
            "get": {
                "description": "List all damage combo cards",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DamageComboCards"
                ],
                "summary": "List all damage combo cards",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.DamageComboCard"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new damage combo card",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DamageComboCards"
                ],
                "summary": "Create a new damage combo card",
                "parameters": [
                    {
                        "description": "Damage Combo Card",
                        "name": "damageComboCard",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/damage_combo_cards.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DamageComboCard"
                        }
                    }
                }
            }
        },
        "/damage_combo_cards/{id}": {
            "get": {
                "description": "Get a damage combo card by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DamageComboCards"
                ],
                "summary": "Get a damage combo card by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Damage Combo Card ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DamageComboCard"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a damage combo card",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DamageComboCards"
                ],
                "summary": "Update a damage combo card",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Damage Combo Card ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Damage Combo Card",
                        "name": "damageComboCard",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/damage_combo_cards.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.DamageComboCard"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a damage combo card",
                "tags": [
                    "DamageComboCards"
                ],
                "summary": "Delete a damage combo card",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Damage Combo Card ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/enemies": {
            "get": {
                "description": "List all enemies",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Enemies"
                ],
                "summary": "List all enemies",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Enemy"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new enemy with enemy moves",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Enemies"
                ],
                "summary": "Create a new enemy",
                "parameters": [
                    {
                        "description": "Enemy",
                        "name": "enemy",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/enemies.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Enemy"
                        }
                    }
                }
            }
        },
        "/enemies/{id}": {
            "get": {
                "description": "Get an enemy by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Enemies"
                ],
                "summary": "Get an enemy by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Enemy ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Enemy"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing enemy",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Enemies"
                ],
                "summary": "Update an existing enemy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Enemy ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Enemy",
                        "name": "enemy",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/enemies.UpdateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Enemy"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an enemy by ID",
                "tags": [
                    "Enemies"
                ],
                "summary": "Delete an enemy by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Enemy ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "damage_combo_cards.CreateRequest": {
            "type": "object",
            "properties": {
                "combo_card_id": {
                    "type": "string"
                },
                "damage": {
                    "type": "integer"
                },
                "range_from": {
                    "type": "integer"
                },
                "range_to": {
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/models.DamageComboCardType"
                }
            }
        },
        "damage_combo_cards.UpdateRequest": {
            "type": "object",
            "properties": {
                "combo_card_id": {
                    "type": "string"
                },
                "damage": {
                    "type": "integer"
                },
                "range_from": {
                    "type": "integer"
                },
                "range_to": {
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/models.DamageComboCardType"
                }
            }
        },
        "enemies.CreateRequest": {
            "type": "object",
            "properties": {
                "class": {
                    "type": "string",
                    "example": "Warrior"
                },
                "defense": {
                    "type": "integer",
                    "example": 2
                },
                "description": {
                    "type": "string",
                    "example": "A small but fierce warrior"
                },
                "enemy_moves": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/enemies.EnemyMoves"
                    }
                },
                "experience": {
                    "type": "integer",
                    "example": 10
                },
                "hp": {
                    "type": "integer",
                    "example": 100
                },
                "level": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Goblin"
                },
                "quantity_deck": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "enemies.EnemyMoves": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Close attack"
                },
                "range_from": {
                    "type": "integer",
                    "example": 1
                },
                "range_to": {
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "enemies.UpdateRequest": {
            "type": "object",
            "properties": {
                "class": {
                    "type": "string"
                },
                "defense": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "experience": {
                    "type": "integer"
                },
                "hp": {
                    "type": "integer"
                },
                "level": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "quantity_deck": {
                    "type": "integer"
                }
            }
        },
...
        "models.DamageComboCard": {
            "type": "object",
            "properties": {
                "combo_card_id": {
                    "type": "string"
                },
                "damage": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "range_from": {
                    "type": "integer"
                },
                "range_to": {
                    "type": "integer"
                },
                "type": {
                    "$ref": "#/definitions/models.DamageComboCardType"
                }
            }
        },
        "models.DamageComboCardType": {
            "type": "integer",
            "enum": [
                0
            ],
            "x-enum-varnames": [
                "PhysicalDamage"
            ]
        },
        "models.Enemy": {
            "type": "object",
            "properties": {
                "class": {
                    "type": "string",
                    "example": "Warrior"
                },
                "defense": {
                    "type": "integer",
                    "example": 2
                },
                "description": {
                    "type": "string",
                    "example": "A small but fierce warrior"
                },
                "enemy_moves": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.EnemyMove"
                    }
                },
                "experience": {
                    "type": "integer",
                    "example": 10
                },
                "hp": {
                    "type": "integer",
                    "example": 100
                },
                "id": {
                    "type": "string"
                },
                "level": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "Goblin"
                },
                "quantity_deck": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "models.EnemyMove": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Slash attack"
                },
                "enemy_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "range_from": {
                    "type": "integer",
                    "example": 1
                },
                "range_to": {
                    "type": "integer",
                    "example": 3
                }
            }
        },
...
    }
}